requirements.txt
picamera
azure-storage-blob
python-dotenv
psutil
retry